Makes ion-hash a feature in ion-rs
