+++ .github/workflows/rust.yml
@@ -44,11 +44,16 @@ jobs:
         with:
           command: build
           args: --verbose --workspace
-      - name: Cargo Test
+      - name: Cargo Test (default/no features)
         uses: actions-rs/cargo@v1
         with:
           command: test
           args: --verbose --workspace
+      - name: Cargo Test (all features)
+        uses: actions-rs/cargo@v1
+        with:
+          command: test
+          args: --verbose --workspace --all-features
       - name: Rustfmt Check
         uses: actions-rs/cargo@v1
         with:
