+++ ion-hash/README.md
@@ -1,19 +1,7 @@
 # `ion-hash-rust`
 
-[![Crate](https://img.shields.io/crates/v/ion-hash.svg)](https://crates.io/crates/ion-hash)
-[![Docs](https://docs.rs/ion-hash/badge.svg)](https://docs.rs/ion-hash)
-[![License](https://img.shields.io/crates/l/ion-hash)](https://crates.io/crates/ion-hash)
-[![CI Build](https://github.com/amazon-ion/ion-rust/workflows/CI%20Build/badge.svg)](https://github.com/amazon-ion/ion-rust/actions?query=workflow%3A%22CI+Build%22)
+***This package is deprecated in favor of the `ion-hash` feature in `ion-rs >= 0.16.0`.***
 
-A Rust implementation of [Ion Hash][spec].
-
-***This package is considered experimental, under active/early development, and the API is subject to change.***
-
-## Development
-
-See [Ion Rust][ion-rust] for details.  This crate is currently developed in concert with Ion Rust
-as a [Cargo workspace][cargo-workspace].
+See [Ion Rust][ion-rust] for details.
 
 [ion-rust]: https://github.com/amazon-ion/ion-rust
-[spec]: https://amazon-ion.github.io/ion-hash/docs/spec.html
-[cargo-workspace]: https://doc.rust-lang.org/cargo/reference/workspaces.html
